Add builders and executable for aggregating browser tests to run with dart2js
This approach is helpful for projects that want to verify their code works in dart2js by running their unit tests in release mode. By generating an aggregate test entrypoint that imports and runs all of the other tests, you can run tests while only having to compile the aggregate entrypoints with dart2js instead of every test. In projects with a lot of tests, this can save a considerable amount of time. For context, this isn’t necessary in the default dev mode because DDC is a modular compiler that can share work, whereas dart2js treats each entrypoint as an isolated program.